+++ internal/provider/dashboards/dashboard_widgets/common.go
@@ -295,6 +295,7 @@ var (
 		"bar_chart",
 		"horizontal_bar_chart",
 		"markdown",
+		"dynamic",
 	}
 )
 
@@ -433,6 +434,76 @@ type WidgetDefinitionModel struct {
 	BarChart           *BarChartModel           `tfsdk:"bar_chart"`
 	HorizontalBarChart *HorizontalBarChartModel `tfsdk:"horizontal_bar_chart"`
 	Markdown           *MarkdownModel           `tfsdk:"markdown"`
+	Dynamic            *DynamicModel            `tfsdk:"dynamic"`
+}
+
+type DynamicModel struct {
+	QueryDefinitions types.List                 `tfsdk:"query_definitions"` //DynamicQueryDefinitionModel
+	TimeFrame        *TimeFrameModel            `tfsdk:"time_frame"`
+	Visualization    *DynamicVisualizationModel `tfsdk:"visualization"`
+}
+
+type DynamicQueryDefinitionModel struct {
+	ID    types.String       `tfsdk:"id"`
+	Name  types.String       `tfsdk:"name"`
+	Query *DynamicQueryModel `tfsdk:"query"`
+}
+
+type DynamicQueryModel struct {
+	Logs      *DynamicQueryLogsModel      `tfsdk:"logs"`
+	Spans     *DynamicQuerySpansModel     `tfsdk:"spans"`
+	Metrics   *DynamicQueryMetricsModel   `tfsdk:"metrics"`
+	DataPrime *DynamicQueryDataPrimeModel `tfsdk:"data_prime"`
+}
+
+type DynamicQueryLogsModel struct {
+	LuceneQuery  types.String `tfsdk:"lucene_query"`
+	GroupBy      types.List   `tfsdk:"group_by"`     //ObservationFieldModel
+	Aggregations types.List   `tfsdk:"aggregations"` //LogsAggregationModel
+	Filters      types.List   `tfsdk:"filters"`      //LogsFilterModel
+	DataModeType types.String `tfsdk:"data_mode_type"`
+}
+
+type DynamicQuerySpansModel struct {
+	LuceneQuery  types.String `tfsdk:"lucene_query"`
+	GroupBy      types.List   `tfsdk:"group_by"`     //SpanObservationFieldModel
+	Aggregations types.List   `tfsdk:"aggregations"` //LogsAggregationModel
+	Filters      types.List   `tfsdk:"filters"`      //SpansFilterModel
+	DataModeType types.String `tfsdk:"data_mode_type"`
+}
+
+type DynamicQueryMetricsModel struct {
+	PromqlQuery     types.String `tfsdk:"promql_query"`
+	PromqlQueryType types.String `tfsdk:"promql_query_type"`
+}
+
+type DynamicQueryDataPrimeModel struct {
+	Query        types.String `tfsdk:"query"`
+	DataModeType types.String `tfsdk:"data_mode_type"`
+}
+
+type SpanObservationFieldModel struct {
+	Keypath      types.List   `tfsdk:"keypath"` //types.String
+	Scope        types.String `tfsdk:"scope"`
+	RelationType types.String `tfsdk:"relation_type"`
+}
+
+type DynamicVisualizationModel struct {
+	Stat *DynamicStatModel `tfsdk:"stat"`
+}
+
+type DynamicStatModel struct {
+	ValueField    types.Object `tfsdk:"value_field"` //ObservationFieldModel
+	Unit          types.String `tfsdk:"unit"`
+	Thresholds    types.List   `tfsdk:"thresholds"` //DynamicThresholdModel
+	ThresholdType types.String `tfsdk:"threshold_type"`
+	Legend        *LegendModel `tfsdk:"legend"`
+}
+
+type DynamicThresholdModel struct {
+	From  types.Float64 `tfsdk:"from"`
+	Color types.String  `tfsdk:"color"`
+	Label types.String  `tfsdk:"label"`
 }
 
 type LineChartModel struct {